backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# LLM and AI
openai==1.3.7
anthropic>=0.40.0
tiktoken==0.5.2

# RAG and embeddings
chromadb==0.4.18
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.3.1
huggingface-hub==0.20.0

# Vector operations
numpy==1.26.2
faiss-cpu==1.7.4

# Visualization
graphviz==0.20.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1

# Data processing
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and monitoring
loguru==0.7.2

# Async database driver
asyncpg